[2.6.38] Fix/investigation for Ubuntu nfs issue.
There's a configuration option which puts the nfs vfstable into read-only memory, when it's compiled as a module.
I haven't worked out what the option is, but it's on by default is Ubuntu 11.04, so we need to handle it.
Map a writable version when we need to write, if the original write would fault.

Talpa: 0.9.76

[git-p4: depot-paths = "//dev/talpa/main/": change = 2095276]