[DEF95174] Fix kernel panic if calling mount with a NULL device path.
It is possible for userspace to call mount() with a NULL device path, which the kernel handles ok (valid for e.g. procfs).
We used to handle this by getting an -EFAULT from the getname() call (not able to access 0).
However we needed to change this behaviour due to the struct filename changes (getname not suitable for generic userspace to kernel-space string copy), and we changed the NULL behaviour to replicate the NULL instead of failing at this point.
We don't need to do the pre-mount scan, as that is for boot-sectors and there needs to be a device for there to be a boot-sector.
Returning -EFAULT allows the mount to complete successfully.

[+DLCL]

[git-p4: depot-paths = "//dev/talpa/main/": change = 2391651]