[LINUXEP-2115] Fix race condition introduced in change #289077
If a filesytem's patch is not prepared and on the patch list when the hook is called, it will fail. So, for a hooked operation to work, we must initialise the patch in this order:

1. prepareFilesystem()
2. add to patch list (GL_object.mPatches)
3. patchFilesystem()

The breaking change had swapped steps 2 and 3, so the patch was being applied before the original address was available to the hooking function.

[git-p4: depot-paths = "//dev/talpa/main/": change = 2581055]